requirements.txt
python-decouple
six
djangorestframework
markdown
django-filter
requests
pycrypto
pyjwt
geoip2
geopy